tustus_2.7.6/requirements.txt
python-telegram-bot[job-queue]==21.6
requests
beautifulsoup4
lxml
python-dotenv
